third_party/blink/web_tests/svg/dynamic-updates/SVGFETurbulenceElement-dom-seed-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'seed' attribute of the SVGFETurbulenceElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS turbulence.getAttribute('seed') is "10"
PASS turbulence.getAttribute('seed') is "5"
PASS successfullyParsed is true

TEST COMPLETE
